% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elcf4R-package.R
\docType{package}
\name{elcf4R-package}
\alias{elcf4R}
\alias{elcf4R-package}
\title{Forecasting Individual Electricity Load Curves}
\description{
\code{elcf4R} provides methods and supporting workflows for day-ahead forecasting
of individual electricity load curves. The current package surface includes
Kernel Wavelet Functional models, clustered KWF, GAM, MARS and LSTM
estimators, dataset adapters for iFlex, StoreNet, Low Carbon London and
REFIT, helpers to build daily segments, and rolling-origin benchmarking
utilities.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://fbertran.github.io/eclf4R/}
  \item \url{https://github.com/fbertran/eclf4R}
  \item Report bugs at \url{https://github.com/fbertran/eclf4R/issues}
}

}
\author{
\strong{Maintainer}: Frederic Bertrand \email{frederic.bertrand@lecnam.net} (\href{https://orcid.org/0000-0002-0837-8281}{ORCID})

Authors:
\itemize{
  \item Fatima Fahs \email{fatima.fahs@es.fr}
  \item Myriam Maumy-Bertrand \email{myriam.maumy@ehesp.fr} (\href{https://orcid.org/0000-0002-4615-1512}{ORCID})
}

}
