% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elcf4r_kwf_clusters.R
\name{elcf4r_assign_kwf_clusters}
\alias{elcf4r_assign_kwf_clusters}
\title{Assign segments to a fitted KWF clustering model}
\usage{
elcf4r_assign_kwf_clusters(object, segments)
}
\arguments{
\item{object}{An \code{elcf4r_kwf_clusters} object.}

\item{segments}{Matrix or data frame of daily segments.}
}
\value{
A character vector of cluster labels.
}
\description{
Assign segments to a fitted KWF clustering model
}
