% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_datasets.R
\name{elcf4r_download_elmas}
\alias{elcf4r_download_elmas}
\title{Download the ELMAS dataset from figshare}
\usage{
elcf4r_download_elmas(dest_dir)
}
\arguments{
\item{dest_dir}{Directory where the files should be unpacked.}
}
\value{
A character vector with the paths of the extracted files.
}
\description{
This function downloads the original ELMAS archive from its public
figshare URL and unpacks it to a local directory.
}
