% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elcf4r_fit_gam.R
\name{elcf4r_fit_gam}
\alias{elcf4r_fit_gam}
\title{Fit a GAM model for load curves}
\usage{
elcf4r_fit_gam(data, use_temperature = FALSE)
}
\arguments{
\item{data}{Data frame with columns \code{y} (load), \code{time_index}
(numeric or factor for within day position), \code{dow}, \code{month},
optional \code{temp} and other covariates.}

\item{use_temperature}{Logical. If \code{TRUE}, include temperature
as smooth effect and interactions.}
}
\value{
An object of class \code{elcf4r_model} with \code{method = "gam"}.
}
\description{
Fit a GAM model for load curves
}
\examples{
id1 <- subset(
  elcf4r_iflex_example,
  entity_id == unique(elcf4r_iflex_example$entity_id)[1]
)
train_data <- subset(id1, date < sort(unique(id1$date))[11])
test_data <- subset(id1, date == sort(unique(id1$date))[11])
fit <- elcf4r_fit_gam(train_data[, c("y", "time_index", "dow", "month", "temp")], TRUE)
pred <- predict(fit, newdata = test_data[, c("y", "time_index", "dow", "month", "temp")])
length(pred)
}
