% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elcf4R-package.R
\docType{package}
\name{elcf4R-package}
\alias{elcf4R}
\alias{elcf4R-package}
\title{elcf4R-package}
\description{
Implements Kernel Wavelet Functional (KWF), Generalized Additive Models (GAM), Multivariate Adaptive Regression Splines (MARS) and RNN LSTM models to forecast individual electricity load curves, following the methodology described in Fahs (2023) and related articles and posters. Includes small demo datasets and helper functions to download large public datasets such as ELMAS, StoreNet Ireland and IDEAL.
}
\examples{
# TODO

}
\references{
TODO
}
\seealso{
TODO
}
\author{
\strong{Maintainer}: Frederic Bertrand \email{frederic.bertrand@lecnam.net} (\href{https://orcid.org/0000-0002-0837-8281}{ORCID})

Authors:
\itemize{
  \item Fatima Fahs \email{fatima.fahs@es.fr}
  \item Myriam Maumy-Bertrand \email{myriam.maumy@ehesp.fr} (\href{https://orcid.org/0000-0002-4615-1512}{ORCID})
}

}
