% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_datasets.R
\name{elcf4r_download_storenet}
\alias{elcf4r_download_storenet}
\title{Download the StoreNet Ireland dataset}
\usage{
elcf4r_download_storenet(dest_dir = tempdir())
}
\arguments{
\item{dest_dir}{Directory where the files should be unpacked.
Defaults to a temporary directory.}
}
\value{
Paths to extracted files.
}
\description{
This function downloads the StoreNet electrical load profiles for
an energy community in Ireland from the authors repository.
}
